requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
Werkzeug==2.2.2
flask-cors
weaviate-client==3.26.0
psycopg2==2.9.9
SQLAlchemy==2.0.24
python-dotenv==1.0.0